+++ doc/sphinx/source/releasenotes.md
@@ -8,6 +8,8 @@ On this page we provide a summary of the main API changes, new features and exam
 
 ### Interface changes
 
+- Add `bool` field type for `CeedQFunctionContext` and related interfaces to use `bool` fields.
+
 ### New features
 
 ### Examples